Arrays, String Support for LLVM Only
Initial cut of LLVM string and array support. Needs some clean up but it's extremely promising.
I also added the ability to compile multiple files as one. Right now that only means that it auto includes the MJC runtime file but that alone makes it so that the small testcases can already be more powerful than they once were.