Added Failing GC Tests
Added GC tests from homework assignment. They seem to fail to reasons related to the GC. I restructured a bit of the GC code and gave a hook to mark global roots for static vars (which I beleive to be the most likely problem). However, even adjusting the heap size to a larger value is causing other existing GC tests to fail.