Interfaces, LLVM Library
Added interfaces to the MJC. They are a subclass of ClassType and play by fairly similar rules. The only thing they don't support of standard Java is default methods, this could probably be added pretty easily but is left out for now.
Added a Class type to the runtime library. Currently it just gives the ability to do instanceof style checks. This didn't turn out to be neccesary and is now latent.
LLVM is now a submodule. Hit some strange memory corruptions when using 3.4, 3.5 libraries and now we are using the Debug+Assertions build to finish our project.
Some code cleanup in the Mjc.jacc makes it easier to handle lists.