GC Write Barrier
Added a basic gc write barrier. When writes are being made to fields that are references, we notify the GC. If the GC determines that the field is in the heap and relevant, it records that for later use. This list is pruned on minor collections and nuked on major collections.